fix(editor3): Delete from reducers the unused code and move a non-reducer function on helpers
