+++ package.json
@@ -132,6 +132,7 @@
   },
   "scripts": {
     "test": "grunt ci:travis",
+    "unit": "grunt unit",
     "lint": "eslint --ext .js --ext .jsx scripts spec tasks *.js",
     "start-test-server": "cd test-server && python3 -m venv env && . env/bin/activate && pip install -Ur requirements.txt && honcho start",
     "protractor": "protractor protractor.conf.js",
