--- scripts/core/editor3/reducers/highlights/index.js
@@ -1,1 +1,0 @@
-export {removeInlineStyles} from './styles';
